+++ lib/haytni/plugins/lockable_plugin.ex
@@ -79,8 +79,7 @@ defmodule Haytni.LockablePlugin do
   end
 
   @impl Haytni.Plugin
-  def files_to_install(_base_path, web_path, scope) do
-    import Mix.Tasks.Haytni.Install, only: [timestamp: 0]
+  def files_to_install(_base_path, web_path, scope, timestamp) do
     [
       # HTML
       {:eex, "views/unlock_view.ex", Path.join([web_path, "views", "haytni", scope, "unlock_view.ex"])},
@@ -90,7 +89,7 @@ defmodule Haytni.LockablePlugin do
       {:eex, "templates/email/lockable/unlock_instructions.text.eex", Path.join([web_path, "templates", "haytni", scope, "email", "lockable", "unlock_instructions.text.eex"])},
       {:eex, "templates/email/lockable/unlock_instructions.html.eex", Path.join([web_path, "templates", "haytni", scope, "email", "lockable", "unlock_instructions.html.eex"])},
       # migration
-      {:eex, "migrations/0-lockable_changes.ex", Path.join([web_path, "..", "..", "priv", "repo", "migrations", "#{timestamp()}_haytni_lockable_#{scope}_changes.ex"])}, # TODO: less "hacky"
+      {:eex, "migrations/0-lockable_changes.ex", Path.join([web_path, "..", "..", "priv", "repo", "migrations", "#{timestamp}_haytni_lockable_#{scope}_changes.ex"])},
     ]
   end