+++ lib/haytni/plugins/recoverable_plugin.ex
@@ -49,8 +49,7 @@ defmodule Haytni.RecoverablePlugin do
   end
 
   @impl Haytni.Plugin
-  def files_to_install(_base_path, web_path, scope) do
-    import Mix.Tasks.Haytni.Install, only: [timestamp: 0]
+  def files_to_install(_base_path, web_path, scope, timestamp) do
     [
       # HTML
       {:eex, "views/password_view.ex", Path.join([web_path, "views", "haytni", scope, "password_view.ex"])},
@@ -61,7 +60,7 @@ defmodule Haytni.RecoverablePlugin do
       {:eex, "templates/email/recoverable/reset_password_instructions.text.eex", Path.join([web_path, "templates", "haytni", scope, "email", "recoverable", "reset_password_instructions.text.eex"])},
       {:eex, "templates/email/recoverable/reset_password_instructions.html.eex", Path.join([web_path, "templates", "haytni", scope, "email", "recoverable", "reset_password_instructions.html.eex"])},
       # migration
-      {:eex, "migrations/0-recoverable_changes.ex", Path.join([web_path, "..", "..", "priv", "repo", "migrations", "#{timestamp()}_haytni_recoverable_#{scope}_changes.ex"])}, # TODO: less "hacky"
+      {:eex, "migrations/0-recoverable_changes.ex", Path.join([web_path, "..", "..", "priv", "repo", "migrations", "#{timestamp}_haytni_recoverable_#{scope}_changes.ex"])},
     ]
   end