+++ lib/haytni/plugins/rememberable_plugin.ex
@@ -65,11 +65,10 @@ defmodule Haytni.RememberablePlugin do
   end
 
   @impl Haytni.Plugin
-  def files_to_install(_base_path, web_path, scope) do
-    import Mix.Tasks.Haytni.Install, only: [timestamp: 0]
+  def files_to_install(_base_path, web_path, scope, timestamp) do
     [
       # migration
-      {:eex, "migrations/0-rememberable_changes.ex", Path.join([web_path, "..", "..", "priv", "repo", "migrations", "#{timestamp()}_haytni_rememberable_#{scope}_changes.ex"])}, # TODO: less "hacky"
+      {:eex, "migrations/0-rememberable_changes.ex", Path.join([web_path, "..", "..", "priv", "repo", "migrations", "#{timestamp}_haytni_rememberable_#{scope}_changes.ex"])},
     ]
   end