+++ lib/haytni/plugins/trackable_plugin.ex
@@ -22,11 +22,10 @@ defmodule Haytni.TrackablePlugin do
   use Haytni.Plugin
 
   @impl Haytni.Plugin
-  def files_to_install(_base_path, web_path, scope) do
-    import Mix.Tasks.Haytni.Install, only: [timestamp: 0]
+  def files_to_install(_base_path, web_path, scope, timestamp) do
     [
       # migration
-      {:eex, "migrations/0-trackable_changes.ex", Path.join([web_path, "..", "..", "priv", "repo", "migrations", "#{timestamp()}_haytni_trackable_#{scope}_changes.ex"])}, # TODO: less "hacky"
+      {:eex, "migrations/0-trackable_changes.ex", Path.join([web_path, "..", "..", "priv", "repo", "migrations", "#{timestamp}_haytni_trackable_#{scope}_changes.ex"])},
     ]
   end