+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - [Authenticable] DELETE method can be overriden for logout by giving the option `logout_method: :get` to your YourApp.Haytni.routes/1 call
 - paths used to generate the routes created by plugins can be customized at your YourApp.Haytni.routes/1 call, see their respective documentation for further details
 - routes created for Haytni are now prefixed by `haytni_<scope>_` to avoid conflicts and permit the use of several Haytni stacks (note that prefixes herited from Phoenix.Router.scope or outer Phoenix.Router.resources still apply)
-- `files_to_install/0` becomes `files_to_install/3` to receive the base_path (the lib/your_app directory), web_path (the lib/your_app_web directory) and scope
+- `files_to_install/0` becomes `files_to_install/4` to receive (in that order) the base_path (the lib/your_app directory), web_path (the lib/your_app_web directory), the scope and a timestamp (included in migration filenames)
 - [Registerable] fix installation of edit.html.eex template, it was simply copied as is instead being evaluated as an EEx template like the others
 
 Upgrade notes: